Alteracoes no Projeto

+++ src/main/java/interfaces/Recepcionar.java
@@ -4,7 +4,7 @@ public interface Recepcionar {
 
     public void darBoasVindas(String nomeHospede);
 
-    public Float cobrarHospedagem(Float valorHospedagem, Float valorDado);
+    public Integer cobrarHospedagem(Integer valorHospedagem, Integer valorDado);
